Consolidated helper methods in one utils.py file and updated imports
